requirements.txt
Django
djangorestframework
scikit-learn
pandas
joblib
